#1374 rename context to adapters
