+++ app/trends/rich-adapters/.classpath
