--- app/trends/rich-context-menu/pom.xml
@@ -1,26 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.phoebus</groupId>
-    <artifactId>app-trends</artifactId>
-    <version>4.6.5-SNAPSHOT</version>
-  </parent>
-  <artifactId>app-trends-rich-context-menu</artifactId>
-  <dependencies>
-    <dependency>
-      <groupId>org.phoebus</groupId>
-      <artifactId>core-ui</artifactId>
-      <version>4.6.5-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.phoebus</groupId>
-      <artifactId>app-databrowser</artifactId>
-      <version>4.6.5-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.phoebus</groupId>
-      <artifactId>app-email-ui</artifactId>
-      <version>4.6.5-SNAPSHOT</version>
-    </dependency>
-  </dependencies>
-</project>
