--- app/trends/rich-context-menu/src/main/resources/META-INF/services/org.phoebus.framework.adapter.AdapterFactory
@@ -1,1 +1,0 @@
-org.phoebus.apps.trends.rich.context.DatabrowserAdapterFactory
