+++ app/trends/simple-adapters/.classpath
