+++ app/trends/simple-adapters/build.xml
@@ -1,0 +1,21 @@
+<project default="app-trends-simple-adapters">
+  <import file="../../../dependencies/ant_settings.xml"/>
+
+  <target name="app-trends-simple-adapters">
+    <mkdir dir="${classes}"/>
+    <javac destdir="${classes}" debug="${debug}">
+      <src path="${src}"/>
+      <classpath>
+        <path refid="app-classpath"/>
+        <pathelement path="../../databrowser/${build}/app-databrowser-${version}.jar"/>
+        <pathelement path="../../email/${build}/app-email-${version}.jar"/>
+      </classpath>
+    </javac>
+  	
+    <jar destfile="${build}/app-trends-simple-adapters-${version}.jar">
+      <fileset dir="${classes}"/>
+      <fileset dir="${resources}"/>
+    </jar>
+  </target>
+	
+</project>
