+++ app/trends/simple-adapters/src/main/java/org/phoebus/apps/trends/simple/adapters/Messages.java
@@ -1,0 +1,22 @@
+package org.phoebus.apps.trends.simple.adapters;
+
+import org.phoebus.framework.nls.NLS;
+
+/**
+ * 
+ * @author Kunal Shroff
+ */
+public class Messages {
+
+    public static String ActionEmailTitle;
+    public static String ActionEmailBody;
+
+    static
+    {
+        NLS.initializeMessages(Messages.class);
+    }
+
+    private Messages()
+    {
+    }
+}
