+++ app/trends/simple-adapters/src/main/resources/META-INF/services/org.phoebus.framework.adapter.AdapterFactory
@@ -1,0 +1,1 @@
+org.phoebus.apps.trends.simple.adapters.DatabrowserAdapterFactory
