+++ app/trends/simple-adapters/src/main/resources/org/phoebus/apps/trends/simple/context/messages.properties
