--- app/trends/simple-context-menu/src/main/java/org/phoebus/apps/trends/simple/context/Messages.java
@@ -1,22 +1,0 @@
-package org.phoebus.apps.trends.simple.context;
-
-import org.phoebus.framework.nls.NLS;
-
-/**
- * 
- * @author Kunal Shroff
- */
-public class Messages {
-
-    public static String ActionEmailTitle;
-    public static String ActionEmailBody;
-
-    static
-    {
-        NLS.initializeMessages(Messages.class);
-    }
-
-    private Messages()
-    {
-    }
-}
