+++ phoebus-product/.classpath
@@ -50,6 +50,6 @@
     <classpathentry combineaccessrules="false" kind="src" path="/app-perfmon"/>
     <classpathentry combineaccessrules="false" kind="src" path="/app-pace"/>
     <classpathentry combineaccessrules="false" kind="src" path="/app-diag"/>
-    <classpathentry combineaccessrules="false" kind="src" path="/app-trends-simple-context-menu"/>
+    <classpathentry combineaccessrules="false" kind="src" path="/app-trends-simple-adapters"/>
     <classpathentry kind="output" path="target/classes"/>
 </classpath>
